requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
decorator==5.1.1
Django==4.0.4
django-cors-headers==3.12.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
feedparser==6.0.10
install==1.3.5
PyJWT==2.4.0
pytz==2022.1
self==2020.12.3
sgmllib3k==1.0.0
sqlparse==0.4.2
